+++ CHANGELOG.md
@@ -32,12 +32,14 @@ Resource data will be available with the resource object. This enhancement helps
 - FCOE network
 - Interconnect type
 - Internal link set
+- Logical enclosure
 - Managed SAN
 - SAS interconnect
 - SAS interconnect type
 - SAS logical interconnect
 - Server hardware
 - Server hardware type
+- Server profile template
 - Switch type
 - Uplink set